+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added EphemeralTaskList and accompanying LLM tools
+- Added TellUserTool for displaying messages to the user without expecting a response
+
 ## [0.5.4] - 2025-04-30
 
 ### Fixed